requirements.txt
appdirs==1.4.3
astroid==2.3.3
attrs==19.3.0
autopep8==1.5
backcall==0.1.0
bleach==3.1.0
cupy-cuda101==7.2.0
cycler==0.10.0
decorator==4.4.1
defusedxml==0.6.0
entrypoints==0.3
fastrlock==0.4
h5py==2.10.0
imageio==2.8.0
importlib-metadata==1.5.0
ipykernel==5.1.4
ipython==7.11.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jedi==0.16.0
Jinja2==2.11.0
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.1.0
jupyter-core==4.6.1
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
llvmlite==0.31.0
Mako==1.1.2
MarkupSafe==1.1.1
matplotlib==3.1.2
mccabe==0.6.1
mistune==0.8.4
mpmath==1.1.0
nbconvert==5.6.1
nbformat==5.0.4
nose==1.3.7
notebook==6.0.3
numba==0.48.0
numexpr==2.7.1
numpy==1.18.1
opencv-python==4.2.0.32
opencv-python-headless==4.2.0.32
openslide-python==1.1.1
pandas==0.25.3
pandocfilters==1.4.2
parso==0.6.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
pkg-resources==0.0.0
prometheus-client==0.7.1
prompt-toolkit==3.0.3
ptyprocess==0.6.0
pycodestyle==2.5.0
pycuda==2019.1.2
pyflakes==2.1.1
Pygments==2.5.2
pylint==2.4.4
pyparsing==2.4.6
pyrsistent==0.15.7
python-dateutil==2.8.1
pytools==2020.1
pytorch-ignite==0.3.0
pytz==2019.3
pyzmq==18.1.1
qtconsole==4.6.0
QtPy==1.9.0
scikit-cuda==0.5.3
scikit-learn==0.22.1
scipy==1.4.1
Send2Trash==1.5.0
six==1.14.0
sympy==1.5.1
terminado==0.8.3
testpath==0.4.4
torch==1.4.0
torchvision==0.5.0
tornado==6.0.3
traitlets==4.3.3
typed-ast==1.4.1
wcwidth==0.1.8
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==2.1.0
